upgrade sms-codec version to 0.0.11 (#11)

+++ sms-server-smpp/src/main/java/io/github/protocol/sms/smpp/server/SmppSubmitMockProcessor.java
@@ -73,7 +73,7 @@ public class SmppSubmitMockProcessor implements SmppSubmitSmProcessor {
                 defaultValue,
                 dataCoding,
                 defaultValue,
-                (byte) content.length,
+                (short) content.length,
                 content);
     }
